wip: refactor result module
