+++ src/interface/http/result-to-http.test.ts
@@ -1,6 +1,7 @@
 import { createResultToHttpMapper, type ResultHttpMapping } from '@/interface/http/result-to-http';
 import { HTTP_CREATED, HTTP_INTERNAL_SERVER_ERROR, HTTP_UNAUTHORIZED } from '@/kairos/shared/http/status-code';
-import { err, ok } from '@/kairos/shared/result/result';
+import { err } from '@/kairos/shared/result/err';
+import { ok } from '@/kairos/shared/result/ok';
 import { describe, expect, it } from 'vitest';
 
 describe('Result to HTTP mapper', () => {
