+++ src/interface/http/result-to-http.ts
@@ -1,6 +1,7 @@
 import { type Normalizer, type SerializeOptions } from '@/interface/http/normalizer.type';
 import { HTTP_INTERNAL_SERVER_ERROR, type StatusCode } from '@/kairos/shared/http/status-code';
-import { isErr, type Result } from '@/kairos/shared/result/result';
+import { isErr } from '@/kairos/shared/result/err';
+import { Result } from '@/kairos/shared/result/result.type';
 
 export interface HttpResponsePayload {
   status: StatusCode;
