+++ src/kairos/authentication/infrastructure/repository/find-one-by-email.ts
@@ -1,7 +1,8 @@
 import { type UsersCollection } from '@/framework/mongodb/schema/users-collection-schema';
 import { invalidCredentialsError } from '@/kairos/authentication/domain/errors';
 import { type FindOneByEmail } from '@/kairos/authentication/domain/user-credentials-repository';
-import { err, ok } from '@/kairos/shared/result/result';
+import { err } from '@/kairos/shared/result/err';
+import { ok } from '@/kairos/shared/result/ok';
 
 type CreateFindOneByEmailDependencies = Readonly<{
   usersCollection: UsersCollection;
