+++ src/kairos/workspace/domain/workspace-repository.ts
@@ -1,4 +1,4 @@
-import type { Result } from '@/kairos/shared/result/result';
+import { Result } from '@/kairos/shared/result/result.type';
 import type { WorkspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
 import type { Workspace } from '@/kairos/workspace/domain/workspace';
