+++ src/kairos/workspace/infrastructure/repository/save-workspace.ts
@@ -1,5 +1,6 @@
 import { type WorkspacesCollection } from '@/framework/mongodb/schema/workspaces-collection-schema';
-import { err, ok } from '@/kairos/shared/result/result';
+import { err } from '@/kairos/shared/result/err';
+import { ok } from '@/kairos/shared/result/ok';
 import { workspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
 import { type SaveWorkspace } from '@/kairos/workspace/domain/workspace-repository';
 import { MongoServerError } from 'mongodb';
